requirements.txt
sounddevice
pillow
pystray
pyaudio